+++ mlflow_cratedb/patch/mlflow/__init__.py
@@ -1,5 +1,6 @@
 from mlflow_cratedb.patch.mlflow.db_types import patch_dbtypes
 from mlflow_cratedb.patch.mlflow.db_utils import patch_db_utils
+from mlflow_cratedb.patch.mlflow.jobs.utils import patch_jobs_utils
 from mlflow_cratedb.patch.mlflow.model import polyfill_refresh_after_dml, polyfill_uniqueness_constraints
 from mlflow_cratedb.patch.mlflow.search_utils import patch_search_utils
 from mlflow_cratedb.patch.mlflow.server import patch_run_server
@@ -16,6 +17,7 @@ def patch_mlflow():
     polyfill_refresh_after_dml()
     polyfill_uniqueness_constraints()
     patch_run_server()
+    patch_jobs_utils()
     patch_environment_variables()
     patch_search_utils()
     patch_tracking()